feat(ui): make toast/hotkey into logical components
