+++ invokeai/frontend/web/src/features/gallery/components/HoverableImage.tsx
@@ -6,7 +6,6 @@ import {
   MenuItem,
   MenuList,
   useDisclosure,
-  useToast,
 } from '@chakra-ui/react';
 import { useAppDispatch, useAppSelector } from 'app/store/storeHooks';
 import { imageSelected } from 'features/gallery/store/gallerySlice';
@@ -35,6 +34,7 @@ import { useFeatureStatus } from 'features/system/hooks/useFeatureStatus';
 import { useParameters } from 'features/parameters/hooks/useParameters';
 import { initialImageSelected } from 'features/parameters/store/actions';
 import { requestedImageDeletion } from '../store/actions';
+import { useAppToaster } from 'app/components/Toaster';
 
 export const selector = createSelector(
   [gallerySelector, systemSelector, lightboxSelector, activeTabNameSelector],
@@ -101,7 +101,7 @@ const HoverableImage = memo((props: HoverableImageProps) => {
 
   const [isHovered, setIsHovered] = useState<boolean>(false);
 
-  const toast = useToast();
+  const toaster = useAppToaster();
 
   const { t } = useTranslation();
 
@@ -176,7 +176,7 @@ const HoverableImage = memo((props: HoverableImageProps) => {
       dispatch(setActiveTab('unifiedCanvas'));
     }
 
-    toast({
+    toaster({
       title: t('toast.sentToUnifiedCanvas'),
       status: 'success',
       duration: 2500,
